README.md: delete "Libs:" lists [ci skip]
Despite our "best effort" to keep these up-to-date, they often fall out
of sync with reality. Also their order was not the one expected by the
linker (`ld` specifically), and guaranteeing the correct ones is a hard
task.

Also, since earlier this year, curl-for-win switched to a unified distro
package (a single .zip), making it straightforward to find out all the
static lib dependencies simply by looking at the .a files distributed in
the package. As for the correct order of them, either use llvm's lld,
which isn't sensitive to it like binutils' ld, or list them between the
options `-Wl,--start-group` and `-Wl,--end-group`.

With this, we're re dropping these manual lists.

Ref: https://github.com/curl/curl-for-win/issues/39