python: install pefile as system package on debian/msys2
If losing all installed Python (pip) packages on every .x upgrade was
not enough (on macOS/Homebrew at least), now it became impossible to
use `pip install --user` (e.g. on Debian). The suggested solution is
to create a venv. The planned pip "escape hatch" mechanisn is not yet
available: `no such option: --break-system-packages`

Our workaround is to limit the pip install method to macOS and install
pefile as system package on Debian, and also MSYS2 where it became
available in late 2021. On Debian testing (and MSYS2) the pefile version
is the same as we installed via pip. Also stop tracking pefile version
and revert to the always latest on macOS. This isn't ideal for
cross-platform reproducibility but the actual impact is likely minor.

Will have to re-evaluate if we still need pefile at all and/or implement
the venv method for macOS.

Ref: https://salsa.debian.org/cpython-team/python3/-/blob/master/debian/README.venv
Ref: https://github.com/pypa/pip/issues/11776
Ref: https://github.com/pypa/pip/pull/11780
Ref: https://peps.python.org/pep-0668/
Ref: https://bugs.gentoo.org/895410

Ref: https://packages.debian.org/unstable/python/python3-pefile
Ref: https://github.com/msys2/MINGW-packages/tree/master/mingw-w64-python-pefile