cmake: avoid odd developer warning with ZLIB_ROOT [ci skip]
This, with `libssh2-cmake.sh` and `libssh.sh`:
```
CMake Warning (dev) at src/CMakeLists.txt:269 (find_package):
  Policy CMP0074 is not set: find_package uses <PackageName>_ROOT variables.
  Run "cmake --help-policy CMP0074" for policy details.  Use the cmake_policy
  command to set the policy and suppress this warning.

  CMake variable ZLIB_ROOT is set to:

    /[...]/zlib/x64-ucrt/usr

  For compatibility, CMake is ignoring the variable.
This warning is for project developers.  Use -Wno-dev to suppress it.
```

It's unclear what this means (even after reading the details), what
action it needs and if no action is necessary from users, why does
CMake show it to users? Without `ZLIB_ROOT`, zlib is not found, so the
setting doesn't seem to be ignored either.

Strangely the same `ZLIB_ROOT` config method doesn't trigger this with
`curl-cmake.sh`. But change it there also anyway for consistency.

We suppress these messages with `-Wno-dev`, but I'd prefer not to, since
we're interested in those for curl.