curl 8.0.0
Also:

Policy change for revision numbers:

We use revision numbers to differenciate between builds of the same
curl release. If we rebuild with newer dependencies, a patch, or
changed build configuration, we bump the revision. This number comes
after the curl version number in the downloadable zip filename (and
the top directory in it): `curl-7.85.0_7-win64-mingw.zip`

Before this patch the first revision was "no revision", e.g.:
`curl-7.85.0-win64-mingw.zip`

After this patch the first revision gets the number "1":
`curl-7.85.0_1-win64-mingw.zip`

TL;DR: We now start with revision "1" instead of an empty revision.

Ref: 2e4bef4be821ab85095fcb2bcf86a597a4233f42