_build.sh: add a second llvm-windres workaround [ci skip]
With llvm-15, llvm-windres-15 is present, but the Windows Resource
woes are not over: In this case we're using a suffixed llvm, where
clang is called clang-15. llvm-windres is not expecting that, tries
to locate `clang` without suffix, and fails.

Our workaround is to symlink windres to a path we control, and add
another symlink called `clang` (to `clang-15`) on the same path.
This makes windres find our suffixed clang and use it.

Alternatives: 1. `clang` alias could be put in PATH, but that's more
intrusive. 2. windres allows to pass the name of the preprocessor
we'd like it to use, but I could not make this work. It also needs
manually passing all low-level preprocessor arguments from scratch.
3. fallback to binutils windres. Also could not make it work.

Links:
https://reviews.llvm.org/D100755
https://github.com/llvm/llvm-project/blob/5bcb4c4da99c443fb880d408e5ff4e9b305bbb77/llvm/tools/llvm-rc/llvm-rc.cpp#L137
https://github.com/msys2/MINGW-packages/discussions/8736

/cc @mstorsjo