boringssl: libcrypto strip breakage and workarounds [ci skip]
binutils strip v2.40 cannot process NASM objects inside libcrypto.a and
fails with an error, leaving the lib unchanged:
```
/usr/local/bin/x86_64-w64-mingw32-strip: x64-ucrt/usr/lib/libcrypto.a(chacha-x86-win.asm.obj): invalid operation
```

llvm-strip v16.0.0 now corrupts NASM objects in libcrypto.a, so it's
also not an option. Here's the error when running llvm-strip on it a
second time:
```
x86_64-w64-mingw32-strip: error: 'libcrypto.a': SymbolTableIndex out of range
```

Disable both for binutils strip and llvm-strip for now. This makes
libcrypto.a not reproducible.

To not leave it like this, introduce another workaround: Turns out,
binutils strip works without the error above if applied to each object
separately. Add this feature to `_libclean.sh` and use it.

I haven't verified if this truly creates a reproducible libcrypto.a
across OSes. It may also behave differently with other combinations of
LLVM and binutils versions.