zlib.sh: enable `BUILD_SHARED_LIBS=OFF` for zlib [ci skip]
It's ineffective to actually disable building the shared lib as of
zlib v1.3. Enable it anyway in the hope that one day it might be fixed,
e.g. by merging this PR opened in 2018:

https://github.com/madler/zlib/pull/347

This would allow an even less obtrusive workaround for the RC
compilation bug, by allowing to disable building the DLL and thus not
hitting that bug in the first place. All this via CMake and without the
manual, GNU Make-level hackery and manual install phase.