_dl.sh: assume PGP signatures [ci skip]
Stop checking for a PGP armor header and just assume we got
a PGP signature file. (unless we got an SSH signature)

This allows to accept binary PGP signatures. Could not readily find
out if a binary blob is a valid PGP signature. Once there is such
method, we might restore checking for them before use.

Binary signatures are provided for the libpsl downloads.