_build.sh: fix mac autotools to not pick headers from /usr/local
Without this, Homebrew or components are picked up from there instead
of the path we explicitly specify. And only at compile time, not link
time, resulting in confused builds using headers from one build of a
dependency and libs from another. Seen with OpenSSL (quictls).

That said the mac builds produced by autotools and cmake are very
different (even with an identical `-V` output).

More info:

I'm running into the OpenSSL-confusion issue in local tests with
autotools. That is on macOS with Apple's CommandLineTools.

Despite passing the correct `./configure` option, what actually gets
picked up is the Homebrew OpenSSL under `/usr/local`. But not at link
time, where it want to link to the correct, custom OpenSSL. The latter
is with `no-engine`, the Homebrew one is with, so it fails to link.
```
$ ./configure [...] --with-openssl=/curl-for-win/quictls/x64-sys/usr
[...]
   CPPFLAGS:        -DNDEBUG -DHAS_ALPN -DCURL_STATICLIB -isystem /curl-for-win/quictls/x64-sys/usr/include
   LDFLAGS:         -framework CoreFoundation -framework SystemConfiguration -framework CoreFoundation -framework Security -L/curl-for-win/quictls/x64-sys/usr/lib
[...]
/bin/sh ../libtool  --tag=CC   --mode=compile clang -arch x86_64 -DHAVE_CONFIG_H \
      -I../../include -I../lib -I../../lib -DBUILDING_LIBCURL -DCURL_HIDDEN_SYMBOLS \
      -DNDEBUG -DHAS_ALPN -DCURL_STATICLIB \
      -isystem /curl-for-win/quictls/x64-sys/usr/include -fvisibility=hidden \
      -mmacosx-version-min=10.13 -O3 -Qunused-arguments -pedantic -Wall -Wextra [-W...]
      -c -o vtls/libcurl_la-openssl.lo `test -f 'vtls/openssl.c' || echo '../../lib/'`vtls/openssl.c
[...]
. vtls/openssl.h
.. /usr/local/include/openssl/ssl.h
... /usr/local/include/openssl/macros.h
.... /usr/local/include/openssl/opensslconf.h
..... /usr/local/include/openssl/configuration.h
.... /usr/local/include/openssl/opensslv.h
[...]

ld: -undefined dynamic_lookup cannot be used to find '_ENGINE_by_id' in dylib in dyld shared cache for architecture x86_64
clang: error: linker command failed with exit code 1 (use -v to see invocation)
```

Changing `-isystem` to `-I` fixes it, though I don't yet know how to
teach autotools to do that. Another fix is to pass this flag to clang:
`-isysroot "$(xcrun --show-sdk-path)"` (this drop /usr/local from the
list of built-in header search paths)