curl: re-enable system LDAP/LDAPS on macOS
We disabled it partly because in macOS CMake builds, LDAPS did
not enable automatically. We wrongly assumed this also applies
to autotools builds, but it turned out this was an issue with
CMake builds only and fixed here:
  https://github.com/curl/curl/pull/12006 (in curl 8.4.0)

With this fix landed, we can re-enable it in CMake and re-enable
with autotools for all curl versions. With 8.3.0 this will only
enable LDAP, without LDAPS.

The other reason for disabling it was that Apple deprecated the
LDAP API in 10.10.

To avoid extra deprecation warnings coming with LDAP on macOS,
we target OS X 10.9 Mavericks instead of High Sierra.

The build also runs fine with these warning present, but we only
enable LDAP/LDAPS on macOS when targeting 10.9.

Also fix `_OSVER` value for `10.[0-9]` inputs.