curl-cmake.sh: enable UNITY builds
Unless building for cross-make compatibility tests. CMake is the only
build method doing unity builds and these binaries will inherently be
much different than non-unity ones. Do not enable when this is goal.

EXPERIMENTAL. Unity mode has been under works for a while now, but it's
not as thoroughly tested as regular builds. These builds require all
static symbols to be unique throughout the whole curl project, which
may fail in build configurations that are not CI tested and that have
not been manually tested either. Regressions are also possible after
curl updates. I expect this to improve over time.