ngtcp2.sh: force disable LIBEV dependency [ci skip]
Without this, it may be picked up from Homebrew or other system paths.
This makes the project build 'examples' (no option to disable it) and
building examples takes unnecessary time, and may also fail with this:

```
[ 98%] Linking CXX executable bsslclient
ld: library not found for -lngtcp2
```

Seen on macOS while testing an AWS-LC H3 build. (But this seems to be
unrelated to AWS-LC specifically.)