_build.sh: enable arm64e for standard llvm for mac [ci skip]
Also set it in the triplet for both apple clang and llvm.

arm64e is arm64 v8.3e allowing pointer-authentication codegen for
apple-a12 CPU and newer (including M1 and newer).

Ref: https://reviews.llvm.org/D87095

For llvm this isn't reflected in `objdump --arch=all --private-headers`
`cpusubtype` output. It says `ALL` instead of `E`. So maybe this change
is not enough.