_build.sh: keep out arm64e designator from autotools [ci skip]
autotools doesn't recognize `arm64e`:
```
Invalid configuration `arm64e-apple-darwin': machine `arm64e-apple' not recognized
```
Ref: https://github.com/curl/curl-for-win/actions/runs/6530947363/job/17732106776#step:3:4161

So backtrack and pass this value only to `-arch` for clang-apple,
and `CMAKE_OSX_ARCHITECTURES` (we will adjust if that fails, but
so far it seems to be accepted). This also revert passing arm64e
to standard llvm which accepts this, but seems to be ignoring it
anyway (as of llvm 16/17).

Regression from 898aeabc9f7f27e1781a1575efb342bf9c8df5b1