set `_WIN32_WINNT=0x0600` (Vista) for Windows [ci skip]
We did not set an explicit Windows target version previously.

curl (as of v8.4.0) doesn't utilize Windows features beyond Vista,
neither do any of the default dependencies. Meaning: binaries are
bit-by-bit identical when setting or not setting this.

However, it is a good idea to set it explititly to Vista, to ensure we
live up to our compatibility goal even if curl or a dependency decides
to implement features specific to newer Windows versions.

Before this patch targets were defined by toolchain defaults:
- x64, x86: 0x0a00 (Win10)
- arm64: 0x0601 (Win7) (requires Win10 anyway for the CPU)

This also works around the `getaddrinfo` regression found in curl
v8.4.0: https://github.com/curl/curl/pull/12217