fix ngtcp2/nghttp3 reproducibility [ci skip]
`ngtcp2_unreachable` and `nghttp3_unreachable` include `__FILE__`
unconditionally. This leaks the full path of the source file when
building with CMake. This in turn may reveal private information
and breaks reproducibility of cross-system builds.

Our solution is to extend the existing local fix we used for LibreSSL
to all CMake projects. (CMake is more prone this problem because it
ends up using absolute source paths.)

Regressions from 2022:

https://github.com/ngtcp2/ngtcp2/commit/31fa4fe7446a8456fa1cca4c579334861fd5a069
https://github.com/ngtcp2/ngtcp2/pull/585

https://github.com/ngtcp2/nghttp3/commit/e9531116d8396834c5ec55ac8254013643d11610
https://github.com/ngtcp2/nghttp3/pull/93