_build.sh: add TLS-specific local install prefix [ci skip]
To make it easier switching between TLS-backends in local tests.

Before this patch we could build each dependency against _one_ specific
openssl fork. It means that when switching to a different openssl
fork, these dependencies had to be rebuilt or manually swapped out.

After this patch a single dependency can hold builds for any number of
openssl forks next to each other and switch between them automatically.

TODO: In packages depending on openssl forks, update `DESTDIR` to point
to the new `_PKGDIRS` variable (instead of `_PKGDIR`). Then make sure
that in per-package post-processing we refer to the build results via
`_PPS` (instead of `_PP`). Then make sure that any dependents also
look for the package via `_PPS` (instead of `_PP`).

These are the openssl-dependent packages:
- ngtcp2
- libssh2
- libssh
- gsasl

The package depending on them is curl.