curl-cmake.sh: delete incorrect comment [ci skip]
Follow-up to 694237ee7210483eeece7eeae7027ba4d8a71f2b

Also the commit message is incorrect in the above commit. There is
actually a way to pass linker flags with CMake. In this case the
`-v` is passed just like that. But, some linker flags we pass both
via CMAKE_C_CLAGS and linker flags. There was a lot of struggle
with this back then, but I cannot remember why it ended up this way
exactly. Something to revisit in the future to see if there is any
cleanup opportunity here.