libressl: switch to cmake [ci skip]
autotools works, and as with libssh2, it could serve as a test bed for
keeping autotools support alive, but in practice, autotools is already
broken in curl-for-win and needs lots of effort to keep working. So I'm
letting it slip in favour of focusing on CMake where supported.

LibreSSL CMake support got a critical fix in v3.8.1 to repair ASM use
with mingw-w64. Then some more in v3.8.2. Its next release (v3.8.3?)
will fix some more issues in apps and libtls, but we don't ship those
so curl-for-win is unaffected.