curl-cmake.sh: stop using `CMAKE_C_STANDARD_LIBRARIES` [ci skip]
Stop passing a copy of `LIBS` via `CMAKE_C_STANDARD_LIBRARIES`.

Started using this variable in a5e0d01e5ab931b9aaf9350048255a386fdafd62
to avoid a hack for `ld` linker failure. But, this didn't solve the
issue and I've restored the hack in
8f38f1c34ca290d51e32e06b8fa161623c8d4ae0. It means this line no longer
serves a purpose and can be deleted.