README.md: make UCRT link more useful [ci skip]
Instead of linking to the technical explanation of UCRT [1], link to the
page that explains how to obtain the UCRT update package manually on
pre-Windows 10 versions.

AFAIU this update should be installing automatically on all supported
systems (Vista and up), but 1. I haven't tested this 2. some systems
may be missing this package regardless, e.g. due to disabled upgrades
or being offline.

Ref: https://github.com/curl/curl/issues/12291

[1] https://devblogs.microsoft.com/cppblog/introducing-the-universal-crt/