trurl.sh: fix passing `-Wl,-B` option with post v0.9 versions [ci skip]
Pass it also via `LDLIBS`, not just `LDFLAGS`. In trurl `Makefile`
`LDFLAGS` comes last, so it did not have an effect on `LDLIBS`. This
only happened with trurl master, because with trurl 0.9, we pass
`LDLIBS` manually and _last_.

It'd probably be nice to use `LDFLAGS` -> `LDLIBS` order in trurl to
make it smoother for most cases.