rework checking for dependencies [ci skip]
Before this patch a dependency was considered present if its main
directory existed.

In local test envs this resulted in picking up whatever dependencies
previously built, which is not necessarily matching the config we're
setting via `CW_CONFIG`. Build was attempted against these dependencies.
Failure was possible if the dependency wasn't built for the targeted
platform yet.

After this patch we consider a dependency present if it should be
present based on `CW_CONFIG`. Build fails if the dependency is missing
or not built yet for the targeted platform.

The goal is to make it easy to switch between configuration just be
updating `CW_CONFIG` in local test environment.

CI builds are not affected by any of this because they were always
downloading and building only the dependencies that we specified with
`CW_CONFIG`. Then the whole env discarded.