add 'openssl' config option [ci skip]
To build with the non-quictls, mainline OpenSSL without H3 support.

Before this patch we controlled this via the `noh3` option, which
selected OpenSSL and disabled H3 at the same time.

This patch makes it possible to build with an H3-enabled TLS-backend
but without enabling the H3 feature by using the `noh3` option.