allow "subset" builds again [ci skip]
One fallout of 1cb8c5bf6360d2d0a110a9800d03e7c20f618e5b is that it
made it impossible to create builds that use a subset of a
configuration by omitting dependencies missing from the disk. E.g.
a default config without libssh2.

Another caveat is that enabled, but missing (not downloaded)
dependencies may be silently missed by certain builds, and/or it
may also be possible that it gets picked up from the environment.

Solve both by verifying that the dependency does exist before use.

Another variant of these cases is when the dependency does exist
(we downloaded/extracted it) but we haven't built it yet for the
target we're building. I may address this in a future commit.

Existing dependencies can still become inconsistent for many
reason!