add `nounity` option to disable UNITY mode [ci skip]
This affects curl and libssh2.

In some cases slower builds may justify the reduction in libcurl (and
libssh2) size when linking them in static mode.

This affects the curl tool, static trurl and any apps' sizes linking to
libcurl statically.

When linking a UNITY-built libcurl to an app, the whole libcurl code is
pulled into the app as one unit. In non-UNITY builds only the parts
actually referenced by the app are pulled in by the linker. libcurl has
a fairly intertwined codebase, so most of the code is included anyway,
but not all. The gain is about 120KB for minimal ("pico") libcurl builds
(on Windows). May be higher in more featureful libcurl configurations.

Also: Make `CW_DEV_CROSSMAKE_REPRO` control UNITY mode libssh2 too.