hide non-libcurl symbols in linux/mac libcurl DLLs [ci skip]
ASM functions aren't hidden. Could not find a way to fix that. It might
need some post-build scrubbing or linker option, or pre-link scrubbing.
I'm not sure.

Effect on binary sizes: gcc tends to increase, llvm to decrease. Except
for musl where gcc builds also decreased in size. With gcc the reason
for the increase is due to curl tool sizes, the shared libcurl is
decreasing all around.

This aims to sync linux/mac libcurl dll with Windows, where we limit
symbols exports to the libcurl API. (despite linking all dependencies
to it statically.)