libressl-cmake.sh: use system `strtonum` for newer mac [ci skip]
LibreSSL's CMake doesn't make an attempt to detect this function, so
enable it manually in a case we know it exists.

This is the reverse of d24d0aeeea33308ed1fee5e90f48c84a28c98db7, which
addressed autotools detecting it even when it shouldn't.

The correct solution would be for LibreSSL to stop re-using,
re-implementing and exporting non-namespaced, standard function names.

In case of `strtonum` it causes a stray export from libcurl dylib on
mac, pending investigation and FIXME.