ci: bump to actions/upload-artifact@v4 for test builds [ci skip]
The advantage is the artifacts appearing for download as they are
uploaded, so no longer have to wait for all jobs to finish.

One observed disadvantage is that download speed seems to capped
at 1.5MB/s, whereas it was at least twice of that for v3.

Leave it at v3 for daily builds for now.