mbedtls: delete support [ci skip]
This backend is breaking on almost every 0.1 release.

Its source tarball was never signed, and now it switched to a manual
tarball (still not signed) that also requires bzip2.

It also didn't support TLS 1.3 up until the latest 3.6.0. But that
release also broke with libssh2 due to hiding an API with no
explanation or a release-notes entry. It's just too much to fixup.
libssh2 would also need yet another overhaul for mbedtls 3 to avoid
using the shortcuts the current support is relying upon (but even
support took a lot of effort to put in place.)

On the other hand LibreSSL is already comparably in binary size,
with a complete featureset, ordely releases.

Ref: https://github.com/Mbed-TLS/mbedtls/releases/tag/v3.6.0