curl.sh: add `curltests` option to build curl tests [ci skip]
The tests built are not run, nor included in the distribution. This is
just a step to allow building them in case we want test coverage for
building curl tests.

`BUILD_TESTING=ON` (the default setting in curl) has a small
performance overhead (about 2 seconds in the 'generate' step, on a slow
non-Linux machine). It doesn't actually build the tests automatically.
For that, we must explicitly build the `testdeps` target.

In these builds we keep `BUILD_TESTING=OFF` by default for maximum
build performance. The `curltests` option makes this `ON` and also
builds the `testdeps` target.