_dl.sh: stop tracking the googlesource.com tarball checksum issue [ci skip]
It's been first reported in 2015-12-13, according to:
https://github.com/google/gitiles/issues/84

The issue has been closed with no resolution. The next issue was still
open when Google archived the whole repository. I don't know where it's
tracked now, but it seems unlikely to ever see a fix.

Keep using the GitHub mirror. Since BoringSSL builds are also broken
now, this option may be deleted in the future.