TODO: omit external CA bundle from distro? [ci skip]
The CA bundle is about to be shipped embedded in the curl tool starting
the next release. It can be extracted with the `--dump-ca-embed` option.

This makes it redundant to ship it as an external file. Esp. for Linux
and macOS builds which don't pick up the external bundle like Windows
does. This may be fixed in the next next release, pending:
https://github.com/curl/curl/pull/14582