_dl.sh: pin BoringSSL, future uncertain [ci skip]
Google seem to have switched to a new release scheme internally for
BoringSSL, and pushes updates to Chrome daily, and/or for each
BoringSSL commit.

A daily chore following these updates downstream is too much to handle.
It also means that using the Chrome repo as an indicator for the latest
stable BoringSSL is no longer workable.

Deciding which commit is safe to use as stable is just as well
unworkable.

(BoringSSL builds stopped building due the hard C++ requirement which
needs extensive work to make it work again in all supported envs.)

How to proceed with BoringSSL, I don't know. Leaving it in the repo
for now as a source of information about such builds and fallouts.