enable c-ares by default in curl 8.12.0 for HTTPS RR support [ci skip]
Replace `noh3` alternative build example with `nocares`.

Size increase is about 200 KiB for the distro `.zip` and `.tar.gz`
files, and about the same amount for `curl.exe` and `libcurl*.dll`.
c-ares static lib is 440 KiB in size (on Windows).

This is not small, but fine. The bigger issue is that c-ares needs
patching to build correctly and it's also unclear if it breaks the
binary on supported systems back to Windows Vista.

[EXPERIMENTAL]

disable c-ares due to extremely slow builds