aws-lc 1.49.1 and pin it [ci skip]
AWS-LC seems to have inherited the churn previously seen with BoringSSL.
Frequent updates are probably nice, but adds a lot of probably low-value
daily maintenance task here to curl-for-win.

Pinning now and bumping on occasion if a new version actually improves
some of the build woes listed in the `boringssl.sh` build script.

Please note that AWS-LC builds, just like BoringSSL ones just crash
on startup, even in case the build happens to work correctly, with
the listed limitations.