curl.sh: drop explicit `crypt32` after upstream fix [ci skip]
Not tested with an actual OpenSSL, but should be unnecessary after an
upstream fix. I forgot why it was necessay before because curl-for-win
builds OpenSSL with `-DUSE_BCRYPTGENRANDOM` and without 'engines' which
is supposed to avoid `crypt32`. Seems to have become a new requirement
with OpenSSL 3.2.0 (Ref: 4ed8fc2ab19a2f1e26ea78906a74725ed77feea3) for
`winstore_store.c` (native CA store support). I can no longer reproduce
the issue with 3.5.0 QUIC API and 8.13.1-DEV curl (before the curl patch
below), likely because this path avoids feature detections in OpenSSL.
With 3.5.0 QUICTLS API it was also unnecessary because of the
pre-fill/fast-track trick used, which avoids feature detection for this
known result.

https://github.com/curl/curl/commit/e325001e523e14ad29c9bf787c85a231d8eae05a
https://github.com/curl/curl/pull/17102