switch to zlib-ng by default
zlib-ng is a modernized, optimized fork of the venerable zlib library.

It builds cleanly, without tricks and hacks. It's an overall smoother
experience, and should also give better performance, though for curl
the difference is expected to be minor.

Build times are a bit higher than classic zlib, but the absolute
difference is negligible.

It makes binaries a 1.5% larger (as tested with Windows 32-bit, libssh):

            |    zlib |  zlibng | increase
:---------- | ------: | ------: | -------:
curl.exe    | 4178432 | 4236800 |    +1.4%
libcurl.dll | 3698176 | 3757056 |    +1.6%