_ci-mac-homebrew.sh: disable WINE crash dialog [ci skip]
Replace forced WINE init with calling the registry tool and disabling
the WINE crash dialog. This also initializas WINE, as done before, and
also ensures that a crashed .exe isn't popping up a GUI window in CI,
causing the job to hang and go gray after timeout:
https://github.com/curl/curl-for-win/actions/runs/16292724921/job/46054493065

Instead, it makes WINE dump crash details to the log and makes the .exe
execution fail as normal, making the CI job red.

Ref: https://gitlab.winehq.org/wine/wine/-/wikis/FAQ#how-do-i-disable-the-gui-crash-dialog