cosign: switch to the new bundle signature format [ci skip]
The signature suffix (`.sigstore`) is subject to change. This is so new,
there is no documentation yet. Existing documentation and its pending
update do not mention or specify a suffix (or I couldn't find it.)
An internet or GitHub search also didn't help.

I've seen so far: .sigstore.json, .bundle, .sig

TODO: also update this in curl/curl-www once decided or before the next
release latest.

It's also an option to stay with the `.cosign` suffix and format, though
not recommended.

Ref: https://github.com/sigstore/cosign/pull/4440
Ref: https://github.com/sigstore/sigstore-blog/pull/89
Ref: https://github.com/sigstore/docs/pull/385