GHA/dependabot: semver cooldown not supported for GHA [ci skip]
Fixing:
```
Dependabot encountered the following error when parsing your .github/dependabot.yml:

The property '#/updates/0/cooldown/semver-major-days' is not supported for the package ecosystem 'github-actions'.
The property '#/updates/0/cooldown/semver-minor-days' is not supported for the package ecosystem 'github-actions'.
The property '#/updates/0/cooldown/semver-patch-days' is not supported for the package ecosystem 'github-actions'.

Please update the config file to conform with Dependabot's specification.
```
Ref: https://github.com/curl/curl-for-win/network/updates