libpsl.sh: set `_FILE_OFFSET_BITS=64`
mingw-w64 13 changed the way `stat()` is mapped in `sys/stat.h`.
This triggered the `stat()` call made from an unused (by curl) libpsl
API to require the `stat64i32` intrinsic symbol, while previously it
mapped to `__imp__stat64`.

This was reported to cause issues with mingw-w64 <13, when linking
8.16.0_4 curl-for-win libraries statically. (8.16.0_3 and earlier were
built with mingw-w64 12, and not affected.)

Try fixing it by setting the macro `_FILE_OFFSET_BITS=64` for the libpsl
build, which makes mingw-w64 map `stat()` to `stat64`.

Reported-by: Morten / Mac-Fly
Bug: https://curl.se/mail/lib-2025-10/0007.html