_ci-mac-homebrew.sh: brew `--quiet` to omit useless warnings [ci skip]
E.g.:
```
jq 1.8.1 is already installed and up-to-date. To reinstall 1.8.1, run: brew reinstall jq
```

These warnings are picked by the GHA workflow page and highlighted as
"Annotations" on the Summary page. This adds noise and require users
to scroll through them to get to the binary artifacts, which are thus
difficult to find.

In practice it seems impossible to create a `brew install` command
that never references already installed packages, as this set changes
all the time in CI, and depends on the runner machine, which may be
a local one with arbitrary packages. Yet we want to ensure that all
packages are present that the build script uses.