psl: add versioned public suffix list updates [ci skip]
Replacing the public suffix list bundled with libpsl.

The original promise / expectation was that libpsl sees regular updates,
and a psl update with it, but the latest release is soon to be 2 years
old, also meaning the bundled psl is old and in need of an update:

https://github.com/rockdaboot/libpsl/releases/tag/0.21.5 (2024-01-13)

The libpsl tarball hosts this psl hash and commit date:
5db9b65997e3c9230ac4353b01994c2ae9667cb9 2024-01-07T11:25:09Z

(This remains the same in libpsl current master)

This patch make psl (Mozilla's public suffix list) a separate package
dependency. Somewhat similar to cacert.

It's also automatically bumped when older than 90 days. Sadly the psl
project doesn't provide tags or versioned releases, so curl-for-win
picks a specific commit hash, and uses the commit's ISO timestamp as
a version number.

The bumper logic got an update to handle this content-addressed
dependency, i.e. there is no extra hash or signature verification. We
trust the git commit hash to always point to the same content on the
remote server hosted by GitHub.

The new psl used is:
399792921792cacc483219997697443b49b6976f 2025-10-16T17:43:24Z