add `noWINE` option [ci skip]
To reduce the amount of packages downloaded for the build.

Downside is that it makes curl/trurl version information missing from
the log and from the package metadata, and the tools are not run after
build as a basic sanity check.

Azure, after 6 months, is still hanging/super-slow sometimes when
trying to install Debian packages it's hosting:
https://github.com/curl/curl/actions/runs/19765210912/job/56636220574?pr=19746

on Debian with llvm:

Before:
Need to get 418 MB of archives.
After this operation, 2204 MB of additional disk space will be used.
https://github.com/curl/curl-for-win/actions/runs/19631975667/job/56213668586

After:
Need to get 74.9 MB of archives.
After this operation, 311 MB of additional disk space will be used.
https://github.com/curl/curl-for-win/actions/runs/19765747974/job/56637979501

on Debian with gcc:

Before:
Need to get 312 MB of archives.
After this operation, 1654 MB of additional disk space will be used.
https://github.com/curl/curl-for-win/actions/runs/19631975667/job/56213668604

After:
Need to get 148 MB of archives.
After this operation, 810 MB of additional disk space will be used.
https://github.com/curl/curl-for-win/actions/runs/19765747974/job/56637979511