_dl.sh: version sort refs/tags returned by GitHub API [ci skip]
Not sure what happened, this used to work for many years, but no longer.
Last time it worked was 1 month ago, or 1.5 max:

0d1b8a1c01b815b945f742fcc3192343c016f660
1cf3be4b9b3c215c4262a8da66e2f5d6ac00f048

Can't spot a local change that could have caused this. Code was relying
on GitHub returning tags in chronological order. Now it's alpha-sorted.

Fix by version sorting the results.